requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.3
browser-cookie3==0.20.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
lz4==4.4.4
outcome==1.3.0.post0
packaging==24.2
pycparser==2.22
pycryptodomex==3.22.0
PySocks==1.7.1
python-dotenv==1.0.1
pywin32==310
requests==2.32.3
selenium==4.29.0
shadowcopy==0.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
WMI==1.5.1
wsproto==1.2.0